refactor: split site infrastructure modules
